saving

+++ src/main/java/com/zensar/vehiclebreakdown/dao/UserDao.java
@@ -1,20 +1,27 @@
 package com.zensar.vehiclebreakdown.dao;
 
 import java.util.List;
+import java.util.Optional;
 
-import org.springframework.data.jpa.repository.JpaContext;
+import javax.persistence.EntityManager;
+
+//import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.lang.Nullable;
+//import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import com.zensar.vehiclebreakdown.model.User;
 
 @Repository
-public interface UserDao extends CrudRepository<User,Integer> {
+public interface UserDao extends CrudRepository<User,Long> {
 
 	public List<User> findByUsertype(String role);
-
-//	public List<User> findAllById(int id);
+//	 @Nullable
+//	public List<User>  findByuser(int user_id);
+//	@Query("select u from User u where u.user_id = ?1")
+//	User getUserById( int user_id);
+//		
 	
-	// public User findById(int id);
-		
+	
 }
